backend_python/requirements.txt
# Python Backend for BEOS (Blood Emergency Platform)

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.10.0
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
